requirements.txt
python-3.11.5
json
pandas
numpy
nltk
langdetect-1.0.9
stopwordsiso-0.6.1
gensim
scikit-learn
scipy
matplotlib
seaborn
wordclouds
ipympl-0.9.4
tiktoken-0.7.0
openai-1.47.0
torch-2.2.2
huggingface-hub-0.25.1 
safetensors-0.4.5 
sentence-transformers-3.1.1 
tokenizers-0.20.0 
transformers-4.45.1